starter/requirements.txt
alembic==1.6.5
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.0.1
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==2.0.0
gunicorn==20.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.1
pyasn1==0.6.1
pytest==8.3.4
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.18
tomli==2.2.1
urllib3==2.3.0
Werkzeug==2.0.1
